speedup: trigger GraphQL query explicitly in componentDidMount
Flame graph showed that, with the graphql HOC applied, the cache
was being checked on each render to reconstruct the places,
which was adding around 90ms to each render cycle, which made
the render chuggy; actual time to update the canvas was only
around 5ms.

I removed the graphql HOC and replaced with withApollo HOC which
injects a ApolloClient into the props, which can then be used
to explicitly make the query call from within componentDidMount.